backend/requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Azure Services
azure-ai-vision-face>=1.0.0b1
azure-cognitiveservices-speech>=1.34.0
azure-ai-textanalytics>=5.3.0
azure-ai-contentsafety>=1.0.0
azure-ai-formrecognizer>=3.3.0
azure-search-documents>=11.4.0
azure-identity>=1.15.0
azure-cosmos>=4.5.0
azure-core>=1.29.0
azure-storage-blob>=12.19.0
azure-monitor-query>=1.2.0

# AI/ML
openai>=1.12.0
anthropic>=0.18.0
librosa>=0.10.0
numpy>=1.26.0,<2.0.0
scikit-learn>=1.4.0
transformers>=4.36.0
torch>=2.1.0
sentence-transformers>=2.3.0

# Caching & Data
redis>=5.0.0
httpx>=0.26.0
celery>=5.3.0
kombu>=5.3.0

# Auth & Security
python-jose[cryptography]==3.3.0
pyjwt==2.8.0
cryptography==41.0.7
msal==1.26.0

# Payments & Auth (optional - commented to avoid pydantic conflicts)
# stripe==7.4.0
# auth0-python==4.4.1

# Network
aiohttp==3.9.1
websockets==12.0

# Utilities
pydantic-settings==2.1.0
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Monitoring & Logging
prometheus-client==0.19.0
structlog==24.1.0
